+++ archguard-gradle-plugin/src/main/kotlin/com/archguard/gradle/DslSupport.kt
@@ -1,0 +1,10 @@
+package com.archguard.gradle
+
+import groovy.lang.Closure
+
+internal fun <T> configureClosure(closure: Closure<*>, target: T): T {
+    val configured = closure.rehydrate(target, closure.owner, closure.thisObject)
+    configured.resolveStrategy = Closure.DELEGATE_FIRST
+    configured.call()
+    return target
+}